move segment metadata package to correct location and add javadoc
